move trait IOAdapter from server to tftp_proto, where it's actually used
